created abstract class for calc
