--- code/mortgage-calculator/src/main/java/ru/unn/agile/mortgagecalculator/model/PeriodType.java
@@ -1,6 +1,0 @@
-package ru.unn.agile.mortgagecalculator.model;
-
-public enum PeriodType {
-    YEAR,
-    MONTH
-}
