+++ code/mortgage-calculator/src/main/java/ru/unn/agile/mortgagecalculator/model/report/PeriodType.java
@@ -1,0 +1,6 @@
+package ru.unn.agile.mortgagecalculator.model.report;
+
+public enum PeriodType {
+    YEAR,
+    MONTH
+}
